+++ inc/sec.php
@@ -12,6 +12,5 @@ if(!isset($_SESSION['parfumvault'])){
 	header('Location: '.$login);
 	exit;
 }
-$pvOnlineAPI = 'https://online.jbparfum.com/api2.php';
 
 ?>