+++ pages/update_data.php
@@ -518,6 +518,8 @@ if($_GET['ingSupplier'] == 'delete'){
 	return;
 }
 
+
+
 if($_POST['value'] && $_GET['formula'] && $_POST['pk']){
 	$value = mysqli_real_escape_string($conn, $_POST['value']);
 	$formula = mysqli_real_escape_string($conn, $_GET['formula']);
@@ -1045,6 +1047,42 @@ if($_POST['action'] == 'clone' && $_POST['old_ing_name'] && $_POST['ing_id']){
 	
 	return;
 }
+
+
+
+//RENAME INGREDIENT
+if($_POST['action'] == 'rename' && $_POST['old_ing_name'] && $_POST['ing_id']){
+	$ing_id = mysqli_real_escape_string($conn, $_POST['ing_id']);
+
+	$old_ing_name = mysqli_real_escape_string($conn, $_POST['old_ing_name']);
+	$new_ing_name = mysqli_real_escape_string($conn, $_POST['new_ing_name']);
+	if(empty($new_ing_name)){
+		$response['error'] = '<strong>Error: </strong>Please enter a name!';
+		echo json_encode($response);
+		return;
+	}
+	if(mysqli_num_rows(mysqli_query($conn, "SELECT name FROM ingredients WHERE name = '$new_ing_name'"))){
+		$response['error'] = '<strong>Error: </strong>'.$new_ing_name.' already exists!';
+		echo json_encode($response);
+		return;
+	}
+	
+	$sql.=mysqli_query($conn, "UPDATE allergens SET ing = '$new_ing_name' WHERE ing = '$old_ing_name'");
+
+	$sql.=mysqli_query($conn, "UPDATE ingredients SET name = '$new_ing_name' WHERE name = '$old_ing_name' AND id = '$ing_id'");
+	$sql.=mysqli_query($conn, "UPDATE formulas SET ingredient = '$new_ing_name' WHERE ingredient = '$old_ing_name' AND ingredient_id = '$ing_id'");
+
+	if($nID = mysqli_fetch_array(mysqli_query($conn, "SELECT name FROM ingredients WHERE name = '$new_ing_name'"))){
+		
+		$response['success'] = $old_ing_name.' renamed to <a href="/pages/mgmIngredient.php?id='.base64_encode($nID['name']).'" >'.$new_ing_name.'</a>!';
+		echo json_encode($response);
+		return;
+	}
+	
+	return;
+}
+
+
 
 header('Location: /');
 exit;