+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # CHANGELOG
 ### Version 5.2
-Extend API to include allergen status
-Extend API to include solvents status
-
+- Extend API to include allergen status
+- Extend API to include solvents status
+- Ingredient rename added
 
 ### Version 5.1
 - New logo design